+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 **unreleased**
 
 **unreleased**
-- [Feature] Remove the validation of the number of files in BatchWriteDetect operator
+- [Feature] Remove the validation of the number of files in `BatchWriteDetectOperator`
+- [Feature] Rename operator `BatchWriteProcessOperator` to `BatchWriteProcessNdjsonOperator`
+- [Feature] Create new operator `BatchWriteProcessOrcOperator` that write ORC file format
 
 **v0.0.52**
 - [Feature] Allow `BatchWriteDetectOperator` to receive a prefix to process only a few files